:broom: 상수 데이터 분리에 관한 사이드 이팩트
