+++ src/api/auth/apis.ts
@@ -4,6 +4,7 @@ import { AuthResponse, UserInfoType } from "@/types/auth";
 const authAPI = {
   /** 카카오 인가 코드 전송 후 로그인 토큰 받아오기 */
   postKakaoCode: async (code: string) => {
+    console.log('서버로 보내는 code',code)
     const { data } = await baseInstance.post<AuthResponse>(`/users/kakao?code=${code}`, {
       headers: {
         "Content-Type": "application/x-www-form-urlencoded;charset=utf-8",
