+++ src/components/auth/AuthForm.tsx
@@ -15,7 +15,8 @@ import authAPI from "@/api/auth/apis";
 import IMAGE_MAP from "@/constants/image";
 import STORAGE_KEYS from "@/constants/storageKeys";
 import { authReducer, initialState } from "@/store/reducer";
-import { ROUTER_PATHS } from "@/utils/router";
+
+import { ROUTER_PATHS } from "@/constants/routerPath";
 import { useMutation } from "@tanstack/react-query";
 import StyledButton from "../common/StyledButton";
 import Validation from "./Validation";
