+++ src/pages/KaKaoLogin.tsx
@@ -4,7 +4,8 @@ import { useNavigate, useSearchParams } from "react-router-dom";
 import authAPI from "@/api/auth/apis";
 import IMAGE_MAP from "@/constants/image";
 import STORAGE_KEYS from "@/constants/storageKeys";
-import { ROUTER_PATHS } from "@/utils/router";
+
+import { ROUTER_PATHS } from "@/constants/routerPath";
 import { useMutation } from "@tanstack/react-query";
 
 const KakaoLoginPage = () => {
